Added a shortcut for pushing Model subclasses
This allows pushing a Model subclass without using the push method directly. Now instead of 'push MyFoo, origin...', one can use 'MyFoo origin...'.

Ruby interprets the new syntax as an attempt to call an undefined method, which triggers method_missing. A hook was added to method_missing to catch this and forward the call to the push method.