Fixed path-related issues with patching Model::DSL
Apparently the path array used by require can be differently ordered, depending on how, or where, the gem is built. Building locally resulted in one ordering, whereas building the gem on TravisCI tended to result in a different ordering. The local builds were on OS X Yosemite, and the Travis builds used Linux, which may have also contributed.

The issue manifested itself as "NoMethodError: undefined method `extrude' for main:Object", which made it rather hard to track down. Apparently having a lib/model.rb in this gem was the issue. If the model gem was in the path first, then all was well. If the model gem was later in the path, the monkey-patch file would be found first, which caused require to decide that the file had already been loaded. Consequently, require wouldn't bother looking for the file provided by the model gem, which resulted in a large number of missing methods.

Moving the monkey-patch version of model.rb into the lib/model directory solved the problem. It may eventually become prudent to think of a better naming scheme for these files.